+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 3.1.21
+
+* If not using effects, items which added bonuses to spell DC did not always apply. e.g. Moon Sickle.
+* Effect: Cloak of Displacement improvements if using midiqol.
+* Effect: Moon Sickle bonus healing if using midi.
+
 # 3.1.20
 
 * If using the "Items with Spells DnD5e" module, spell data for items will be generated. If "Magic Items" is also installed, the importer will use that first.